refact: added types files, requests and responses folders
