+++ src/entities/index.ts
@@ -1,5 +1,3 @@
-export { IUser } from './user';
+export { IUser, UpdateArgs } from './types';
 
 export { User } from './user';
-
-export { UpdateArgs } from './user';
