+++ src/router/index.ts
@@ -1,3 +1,1 @@
 export { Router } from './router';
-
-export { Request } from './request';
