+++ src/server/types.ts
@@ -1,0 +1,4 @@
+import { ServerResponse } from 'http';
+import { Request } from '../requests';
+
+export type RequestListenerHandler = (req: Request, res: ServerResponse) => void;
