+++ src/services/index.ts
@@ -2,12 +2,4 @@ export { UserService } from './user.services';
 
 export { UserWorkerService } from './user.worker.services';
 
-export { IUserService } from './user.services';
-
-export {
-    httpStatusCodes,
-    ErrorMessages,
-    RequestHeaders,
-    ResponseHandler,
-    ErrorHandler
-} from './response';
+export { IUserService } from './types';
